requirements.txt
asgiref==3.3.4
certifi==2020.12.5
click==7.1.2
colorama==0.4.4
Django==3.2.1
image==1.5.33
numpy==1.19.5
Pillow==8.2.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.4.3.0.3
python-dotenv==0.17.1
pytz==2021.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
QtPy==1.9.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
typing-extensions==3.10.0.0
wincertstore==0.2
